feat: deleteByIds returns number; 'in' query operator

+++ src/simpleFile.db.ts
@@ -1,4 +1,3 @@
-import { StringMap } from '@naturalcycles/js-lib'
 import * as fs from 'fs-extra'
 import { Observable, Subject } from 'rxjs'
 import { BaseDBEntity, CommonDB, CommonDBOptions, CommonDBSaveOptions } from './db.model'
@@ -31,14 +30,18 @@ export class SimpleFileDB implements CommonDB {
 
   cfg!: Required<SimpleFileDBCfg>
 
-  cache: StringMap<StringMap<any>> = {}
+  cache: Record<string, Record<string, any>> = {}
 
-  async resetCache (): Promise<void> {
-    await fs.emptyDir(this.cfg.storageDir)
-    this.cache = {}
+  async resetCache (table?: string): Promise<void> {
+    if (table) {
+      await this.saveTable(table, {})
+    } else {
+      await fs.emptyDir(this.cfg.storageDir)
+      this.cache = {}
+    }
   }
 
-  private async getTable<DBM extends BaseDBEntity> (table: string): Promise<StringMap<DBM>> {
+  private async getTable<DBM extends BaseDBEntity> (table: string): Promise<Record<string, DBM>> {
     if (!this.cache[table]) {
       this.cache[table] = await fs
         .readJson(`${this.cfg.storageDir}/${table}.json`)
@@ -49,7 +52,7 @@ export class SimpleFileDB implements CommonDB {
 
   private async saveTable<DBM extends BaseDBEntity> (
     table: string,
-    data: StringMap<DBM>,
+    data: Record<string, DBM>,
   ): Promise<void> {
     this.cache[table] = data
     const filePath = `${this.cfg.storageDir}/${table}.json`
@@ -66,7 +69,7 @@ export class SimpleFileDB implements CommonDB {
     return ids.map(id => data[id]).filter(Boolean)
   }
 
-  async deleteByIds (table: string, ids: string[], opts?: CommonDBOptions): Promise<string[]> {
+  async deleteByIds (table: string, ids: string[], opts?: CommonDBOptions): Promise<number> {
     const data = await this.getTable(table)
     const deletedIds: string[] = []
     ids.forEach(id => {
@@ -77,7 +80,7 @@ export class SimpleFileDB implements CommonDB {
     })
     await this.saveTable(table, data)
 
-    return deletedIds
+    return deletedIds.length
   }
 
   async saveBatch<DBM extends BaseDBEntity> (
@@ -96,14 +99,14 @@ export class SimpleFileDB implements CommonDB {
     q: DBQuery<DBM>,
     opts?: CommonDBOptions,
   ): Promise<DBM[]> {
-    return queryInMemory(q, await this.getTable(q.table))
+    return queryInMemory(q, Object.values(await this.getTable(q.table)))
   }
 
   async runQueryCount<DBM extends BaseDBEntity> (
     q: DBQuery<DBM>,
     opts?: CommonDBOptions,
   ): Promise<number> {
-    const rows = queryInMemory(q, await this.getTable(q.table))
+    const rows = queryInMemory(q, Object.values(await this.getTable(q.table)))
     return rows.length
   }
 
@@ -111,7 +114,7 @@ export class SimpleFileDB implements CommonDB {
     const subj = new Subject<DBM>()
 
     void this.getTable<DBM>(q.table).then(data => {
-      queryInMemory(q, data).forEach(dbm => subj.next(dbm))
+      queryInMemory(q, Object.values(data)).forEach(dbm => subj.next(dbm))
       subj.complete()
     })
 
@@ -121,12 +124,12 @@ export class SimpleFileDB implements CommonDB {
   async deleteByQuery<DBM extends BaseDBEntity> (
     q: DBQuery<DBM>,
     opts?: CommonDBOptions,
-  ): Promise<string[]> {
+  ): Promise<number> {
     const data = await this.getTable(q.table)
-    const rows = queryInMemory(q, data)
+    const rows = queryInMemory(q, Object.values(data))
     const deletedIds = rows.map(dbm => dbm.id)
     deletedIds.forEach(id => delete data[id])
     await this.saveTable(q.table, data)
-    return deletedIds
+    return deletedIds.length
   }
 }
